V1.5
In this Versión the game is almost done, there´s now the ninja bomb, But be really carefuly with him, if you touch him you´ll lose lifes, and he is really similar to the other one! Also, the notifier system is already completed for you to be easier to do what you want when you finish the level. The restart button is already fixed and now there´s completed all the interface. The game is almost done :D