V1.6
In this version there´s fixed the 3rd level because on the last versions that was not able to winning it, it was impossible unless you are a machine or something like that, the time of the level is changed from 20 seconds to 30 seconds, the number of kills to win the level is changed from 30 kills to 25 kills, and the times of the 1st and 2nd level were changed to 40 and 50 seconds respectively. Consequently of the above, there was added 1 ninja-bomb more to keep the hardly difficulty of the 3rd level.